spaces/img2img-api/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.34.0
requests==2.32.3
Pillow==11.1.0

# ML stack (installed in Dockerfile): torch/torchvision
diffusers==0.32.2
transformers==4.48.3
accelerate==1.3.0
safetensors==0.5.2
huggingface_hub==0.27.1

